+++ components/webhooks/src/messages.js
@@ -1,0 +1,4 @@
+
+module.exports = {
+  BOOT_MESSAGE: 'webhooksServiceBoot',
+};